zensical:fix open browser in a thread
This ensures that the the server is actually started before opening the browser.
Without this, the browser can, and generally does, open to before the server
has started, resulting in a server not found type of error. Th result is the
user has to manually reload the browser on the first load.